BB-2793: Change CombinedPriceListCommand cron to process queue items …

+++ src/OroB2B/Bundle/PricingBundle/Builder/CombinedPriceListActivationPlanBuilder.php
@@ -2,6 +2,7 @@
 
 namespace OroB2B\Bundle\PricingBundle\Builder;
 
+use Doctrine\Common\Persistence\ObjectManager;
 use Oro\Bundle\EntityBundle\ORM\DoctrineHelper;
 
 use OroB2B\Bundle\PricingBundle\Entity\CombinedPriceList;
@@ -54,17 +55,37 @@ class CombinedPriceListActivationPlanBuilder
     protected $CPLActivationRuleRepository;
 
     /**
+     * @var ObjectManager
+     */
+    protected $manager;
+
+    /**
+     * @var array
+     */
+    protected $processedPriceLists = [];
+
+    /**
+     * @var array
+     */
+    protected $processedCPLs = [];
+
+    /**
      * @param DoctrineHelper $doctrineHelper
      * @param PriceListScheduleResolver $schedulerResolver
-     * @param CombinedPriceListProvider $combinedPriceListProvider
      */
     public function __construct(
         DoctrineHelper $doctrineHelper,
-        PriceListScheduleResolver $schedulerResolver,
-        CombinedPriceListProvider $combinedPriceListProvider
+        PriceListScheduleResolver $schedulerResolver
     ) {
         $this->doctrineHelper = $doctrineHelper;
         $this->schedulerResolver = $schedulerResolver;
+    }
+
+    /**
+     * @param CombinedPriceListProvider $combinedPriceListProvider
+     */
+    public function setProvider(CombinedPriceListProvider $combinedPriceListProvider)
+    {
         $this->combinedPriceListProvider = $combinedPriceListProvider;
     }
 
@@ -73,11 +94,15 @@ class CombinedPriceListActivationPlanBuilder
      */
     public function buildByPriceList(PriceList $priceList)
     {
+        if ($this->isPriceListProcessed($priceList)) {
+            return;
+        }
         $cplIterator = $this->getCombinedPriceListRepository()->getCombinedPriceListsByPriceList($priceList);
 
         foreach ($cplIterator as $cpl) {
             $this->buildByCombinedPriceList($cpl);
         }
+        $this->addPriceListProcessed($priceList);
     }
 
     /**
@@ -85,8 +110,12 @@ class CombinedPriceListActivationPlanBuilder
      */
     public function buildByCombinedPriceList(CombinedPriceList $cpl)
     {
+        if ($this->isCPLProcessed($cpl)) {
+            return;
+        }
         $this->getCPLActivationRuleRepository()->deleteRulesByCPL($cpl);
         $this->generateActivationRules($cpl);
+        $this->addCPLProcessed($cpl);
     }
 
     /**
@@ -95,13 +124,11 @@ class CombinedPriceListActivationPlanBuilder
      */
     protected function generateActivationRules(CombinedPriceList $cpl)
     {
-        $priceListSchedules = $this->getPriceListScheduleRepository()->getSchedulesByCPL($cpl);
+        $now = new \DateTime('now', new \DateTimeZone('UTC'));
+        $priceListSchedules = $this->getPriceListScheduleRepository()->getSchedulesByCPL($cpl, $now);
         $priceListRelations = $this->getCPLToPriceListRepository()->getPriceListRelations($cpl);
 
         $rawRules = $this->schedulerResolver->mergeSchedule($priceListSchedules, $priceListRelations);
-        $activationRuleClass = 'OroB2B\Bundle\PricingBundle\Entity\CombinedPriceListActivationRule';
-        $manager = $this->doctrineHelper->getEntityManagerForClass($activationRuleClass);
-        $now = new \DateTime('now', new \DateTimeZone('UTC'));
         foreach ($rawRules as $ruleData) {
             if ($ruleData[PriceListScheduleResolver::EXPIRE_AT_KEY] !== null
                 && $now > $ruleData[PriceListScheduleResolver::EXPIRE_AT_KEY]) {
@@ -116,18 +143,14 @@ class CombinedPriceListActivationPlanBuilder
             if ($ruleData[PriceListScheduleResolver::ACTIVATE_AT_KEY]) {
                 $rule->setActivateAt($ruleData[PriceListScheduleResolver::ACTIVATE_AT_KEY]);
             }
-            if ($ruleData[PriceListScheduleResolver::ACTIVATE_AT_KEY] === null
-                || $ruleData[PriceListScheduleResolver::ACTIVATE_AT_KEY] < $now) {
-                $rule->setActive(true);
-            }
             $actualCPL = $this->combinedActualCombinedPriceList(
                 $priceListRelations,
                 $ruleData[PriceListScheduleResolver::PRICE_LISTS_KEY]
             );
             $rule->setCombinedPriceList($actualCPL);
-            $manager->persist($rule);
+            $this->getManager()->persist($rule);
         }
-        $manager->flush();
+        $this->getManager()->flush();
     }
 
     /**
@@ -146,10 +169,7 @@ class CombinedPriceListActivationPlanBuilder
                 );
             }
         }
-        return $this->combinedPriceListProvider->getCombinedPriceList(
-            $sequence,
-            CombinedPriceListProvider::BEHAVIOR_EMPTY
-        );
+        return $this->combinedPriceListProvider->getCombinedPriceList($sequence);
     }
 
     /**
@@ -201,5 +221,52 @@ class CombinedPriceListActivationPlanBuilder
                 ->getEntityRepository('OroB2BPricingBundle:CombinedPriceListActivationRule');
         }
         return $this->CPLActivationRuleRepository;
+    }
+
+    /**
+     * @return ObjectManager
+     */
+    public function getManager()
+    {
+        if (!$this->manager) {
+            $this->manager = $this->doctrineHelper
+                ->getEntityManagerForClass('OroB2BPricingBundle:CombinedPriceListActivationRule');
+        }
+
+        return $this->manager;
+    }
+
+    /**
+     * @param PriceList $priceList
+     * @return bool
+     */
+    protected function isPriceListProcessed(PriceList $priceList)
+    {
+        return !empty($this->processedPriceLists[$priceList->getId()]);
+    }
+
+    /**
+     * @param CombinedPriceList $cpl
+     * @return bool
+     */
+    protected function isCPLProcessed(CombinedPriceList $cpl)
+    {
+        return !empty($this->processedCPLs[$cpl->getId()]);
+    }
+
+    /**
+     * @param PriceList $priceList
+     */
+    protected function addPriceListProcessed(PriceList $priceList)
+    {
+        $this->processedPriceLists[$priceList->getId()] = true;
+    }
+
+    /**
+     * @param CombinedPriceList $cpl
+     */
+    protected function addCPLProcessed(CombinedPriceList $cpl)
+    {
+        $this->processedCPLs[$cpl->getId()] = true;
     }
 }
